[fix] Bump package version
This repo publish to npm on every new commit that is pushed to main. And npm
requires a new version number on publishing. Previously, developer need to
manually update the version number to make publish success(and sometime the
forget to update it which result in a failure:

e.g. https://github.com/taoning2014/srt-validator/actions/runs/1707734150

This commit automate the process by using pre-commit hook to bump the version.